// lib/types.ts
export type Collection = {
  id: string;
  name: string;
  description: string;
  image: string;
  category: string;
};

export type ServiceItem = {
  id: string;
  title: string;
  description: string;
  icon: string;
};

export type ValueItem = {
  title: string;
  description: string;
};

export type WhyChooseItem = {
  title: string;
  description: string;
};